requirements.txt
requests
python-dotenv
flask
numpy
pandas
scikit-learn